casecore-runtime/production/backend/requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
sqlalchemy>=2.0.23,<3.0.0
pydantic>=2.5.2,<3.0.0
python-multipart>=0.0.6,<1.0.0
websockets>=12.0,<13.0
aiosqlite>=0.19.0,<1.0.0
pyyaml>=6.0,<7.0
